pyproject.toml
[tool.poetry]
name = "ticket"
version = "0.1.0"
description = ""
authors = ["Lwin Maung Maung <lwinmaungmaung.ytu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.0.1"
pylint = "^3.0.4"
uvicorn = "^0.27.1"
starlette = "^0.37.1"
strawberry-graphql = {extras = ["debug-server"], version = "^0.219.2"}
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
pydantic = "^2.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
